Refactor: Split large CLI command files into logical modules
Split monolithic CLI command files into focused modules to improve maintainability and code organization.

## Files Refactored:

### `email.rs` (867 lines → 6 modules)
- Split into: analytics, core, providers, queue, templates, testing
- Maintains backward compatibility with existing API

### `queue.rs` (795 lines → 3 modules)
- Split into: queue_work, queue_status, queue_scheduler
- Separated concerns: work processing, status monitoring, scheduling

### `auth.rs` (761 lines → 3 modules)
- Split into: auth_setup, auth_scaffold, auth_generators
- Organized: setup/config, scaffolding, key generation

### `new.rs` (492 lines → 2 modules)
- Split into: new_app, new_templates
- Separated: app creation logic and template generation

### `interactive_setup.rs` (410 lines → 2 modules)
- Split into: interactive_wizard, interactive_config
- Organized: wizard flow and configuration handling

## Benefits:
- **Improved maintainability**: Each module has focused responsibility
- **Better code organization**: Related functionality grouped together
- **Enhanced readability**: Smaller, more manageable files
- **Zero breaking changes**: All existing APIs preserved
- **Consistent patterns**: Modular structure across all commands

## Testing:
- ✅ All CLI commands function correctly
- ✅ Help system works for all commands
- ✅ Build passes with no errors
- ✅ Module imports and exports work properly
