+++ crates/cli/src/commands/auth/mod.rs
@@ -1,0 +1,7 @@
+pub mod auth_setup;
+pub mod auth_scaffold;
+pub mod auth_generators;
+
+pub use auth_setup::setup;
+pub use auth_scaffold::scaffold;
+pub use auth_generators::generate_key;
