+++ crates/cli/src/commands/email/analytics.rs
@@ -1,0 +1,26 @@
+use elif_core::ElifError;
+
+#[derive(Debug)]
+pub struct EmailTrackAnalyticsArgs {
+    pub range: String,
+    pub filter: Option<String>,
+}
+
+/// Show email tracking analytics
+pub async fn track_analytics(args: EmailTrackAnalyticsArgs) -> Result<(), ElifError> {
+    println!("📊 Email Analytics - Range: {}", args.range);
+    if let Some(filter) = &args.filter {
+        println!("🎯 Filter: {}", filter);
+    }
+    println!("⏳ Analytics not yet implemented");
+    // TODO: Connect to analytics backend and show data
+    Ok(())
+}
+
+/// Show email delivery statistics
+pub async fn track_stats(group_by: &str) -> Result<(), ElifError> {
+    println!("📈 Email Statistics - Grouped by: {}", group_by);
+    println!("⏳ Statistics not yet implemented");
+    // TODO: Connect to analytics backend and show stats
+    Ok(())
+}
