+++ crates/cli/src/commands/email/providers.rs
@@ -1,0 +1,38 @@
+use elif_core::ElifError;
+
+#[derive(Debug)]
+pub struct EmailProviderConfigureArgs {
+    pub provider: String,
+    pub interactive: bool,
+}
+
+/// Test email provider connection
+pub async fn provider_test(provider: Option<String>) -> Result<(), ElifError> {
+    let provider_name = provider.unwrap_or_else(|| "default".to_string());
+    println!("🔌 Testing email provider: {}", provider_name);
+    println!("⏳ Provider testing not yet implemented");
+    // TODO: Load provider config and test connection
+    Ok(())
+}
+
+/// Configure email provider
+pub async fn provider_configure(args: EmailProviderConfigureArgs) -> Result<(), ElifError> {
+    println!("⚙️  Configuring email provider: {}", args.provider);
+    if args.interactive {
+        println!("🎯 Interactive configuration mode");
+        // TODO: Launch interactive configuration wizard
+    } else {
+        println!("📋 Non-interactive configuration");
+        // TODO: Use environment variables or config files
+    }
+    println!("⏳ Provider configuration not yet implemented");
+    Ok(())
+}
+
+/// Switch active email provider
+pub async fn provider_switch(provider: &str) -> Result<(), ElifError> {
+    println!("🔄 Switching to email provider: {}", provider);
+    println!("⏳ Provider switching not yet implemented");
+    // TODO: Update active provider in config
+    Ok(())
+}
