+++ crates/cli/src/commands/email/queue.rs
@@ -1,0 +1,53 @@
+use elif_core::ElifError;
+
+#[derive(Debug)]
+pub struct EmailQueueProcessArgs {
+    pub limit: Option<u32>,
+    pub timeout: u64,
+}
+
+#[derive(Debug)]
+pub struct EmailQueueClearArgs {
+    pub failed: bool,
+    pub completed: bool,
+}
+
+/// Show email queue status
+pub async fn queue_status(detailed: bool) -> Result<(), ElifError> {
+    println!("📊 Email Queue Status:");
+    if detailed {
+        println!("📋 Detailed information enabled");
+    }
+    println!("⏳ Queue status not yet implemented");
+    // TODO: Connect to queue backend and show status
+    Ok(())
+}
+
+/// Process queued emails
+pub async fn queue_process(args: EmailQueueProcessArgs) -> Result<(), ElifError> {
+    println!("⚡ Processing email queue");
+    if let Some(limit) = args.limit {
+        println!("📏 Processing up to {} emails", limit);
+    }
+    println!("⏰ Timeout: {} seconds per email", args.timeout);
+    println!("⏳ Queue processing not yet implemented");
+    // TODO: Connect to queue backend and process emails
+    Ok(())
+}
+
+/// Clear email queue
+pub async fn queue_clear(args: EmailQueueClearArgs) -> Result<(), ElifError> {
+    println!("🧹 Clearing email queue");
+    if args.failed {
+        println!("❌ Clearing failed jobs");
+    }
+    if args.completed {
+        println!("✅ Clearing completed jobs");
+    }
+    if !args.failed && !args.completed {
+        return Err(ElifError::Validation("Must specify either --failed or --completed".to_string()));
+    }
+    println!("⏳ Queue clearing not yet implemented");
+    // TODO: Connect to queue backend and clear specified jobs
+    Ok(())
+}
