+++ crates/cli/src/commands/email/templates.rs
@@ -1,0 +1,44 @@
+use elif_core::ElifError;
+use std::collections::HashMap;
+use serde_json::{Value, from_str};
+
+#[derive(Debug)]
+pub struct EmailTemplateRenderArgs {
+    pub template: String,
+    pub context: Option<String>,
+    pub format: String,
+}
+
+/// List all available email templates
+pub async fn template_list() -> Result<(), ElifError> {
+    println!("📄 Available Email Templates:");
+    println!("⏳ Template discovery not yet implemented");
+    // TODO: Scan templates directory and list available templates
+    Ok(())
+}
+
+/// Validate email template syntax
+pub async fn template_validate(template: &str) -> Result<(), ElifError> {
+    println!("🔍 Validating template: {}", template);
+    println!("⏳ Template validation not yet implemented");
+    // TODO: Load template and validate Tera syntax
+    Ok(())
+}
+
+/// Render email template with context data
+pub async fn template_render(args: EmailTemplateRenderArgs) -> Result<(), ElifError> {
+    println!("🎨 Rendering template: {}", args.template);
+    
+    // Parse context data if provided
+    let _context_data: HashMap<String, Value> = if let Some(context_str) = &args.context {
+        from_str(context_str)
+            .map_err(|e| ElifError::Validation(format!("Invalid JSON context: {}", e)))?
+    } else {
+        HashMap::new()
+    };
+    
+    println!("📋 Output format: {}", args.format);
+    println!("⏳ Template rendering not yet implemented");
+    // TODO: Load template, render with context, and display output
+    Ok(())
+}
