+++ crates/cli/src/commands/new/mod.rs
@@ -1,0 +1,4 @@
+pub mod new_app;
+pub mod new_templates;
+
+pub use new_app::create_app;
