+++ crates/cli/src/commands/queue/queue_scheduler.rs
@@ -5,390 +5,6 @@ use crate::command_system::{CommandHandler, CommandError, CommandDefinition, imp
 use async_trait::async_trait;
 use chrono::Timelike;
 
-/// Queue work command arguments
-#[derive(Args, Debug, Clone)]
-pub struct QueueWorkArgs {
-    /// Queue names to process (comma-separated)
-    #[arg(long, short, default_value = "default")]
-    pub queue: String,
-    
-    /// Maximum number of jobs to process
-    #[arg(long, short)]
-    pub max_jobs: Option<u32>,
-    
-    /// Timeout in seconds for each job
-    #[arg(long, short, default_value = "60")]
-    pub timeout: u64,
-    
-    /// Sleep time between checks (in milliseconds)
-    #[arg(long, default_value = "1000")]
-    pub sleep: u64,
-    
-    /// Number of worker processes
-    #[arg(long, short, default_value = "1")]
-    pub workers: u8,
-    
-    /// Stop after this many seconds
-    #[arg(long)]
-    pub stop_when_empty: bool,
-    
-    /// Show verbose output
-    #[arg(long, short)]
-    pub verbose: bool,
-}
-
-impl_command!(
-    QueueWorkArgs,
-    "queue:work",
-    "Process background jobs from specified queues",
-    "Process background jobs from one or more queues.\n\n\
-     Features:\n\
-     - Multiple queue support\n\
-     - Configurable timeouts and worker count\n\
-     - Graceful shutdown handling\n\
-     - Job retry mechanisms\n\
-     - Verbose logging and monitoring\n\n\
-     Examples:\n\
-       elifrs queue:work --queue high,default\n\
-       elifrs queue:work --workers 4 --timeout 120\n\
-       elifrs queue:work --stop-when-empty --max-jobs 100\n\
-       elifrs queue:work --verbose --sleep 500"
-);
-
-/// Queue work command handler
-pub struct QueueWorkCommand {
-    pub args: QueueWorkArgs,
-}
-
-#[async_trait]
-impl CommandHandler for QueueWorkCommand {
-    async fn handle(&self) -> Result<(), CommandError> {
-        println!("🔄 Starting queue worker...");
-        
-        let queues: Vec<&str> = self.args.queue.split(',').collect();
-        println!("📋 Processing queues: {:?}", queues);
-        println!("👥 Workers: {}", self.args.workers);
-        println!("⏱️  Job timeout: {}s", self.args.timeout);
-        println!("😴 Sleep interval: {}ms", self.args.sleep);
-        
-        if let Some(max) = self.args.max_jobs {
-            println!("🎯 Max jobs: {}", max);
-        }
-        
-        if self.args.stop_when_empty {
-            println!("🛑 Will stop when queues are empty");
-        }
-        
-        if self.args.verbose {
-            println!("🔊 Verbose mode enabled");
-        }
-        
-        // Check if we have a database connection configured
-        self.verify_database_setup().await?;
-        
-        // Initialize job processing
-        self.start_workers(queues).await?;
-        
-        Ok(())
-    }
-    
-    fn name(&self) -> &'static str {
-        QueueWorkArgs::NAME
-    }
-    
-    fn description(&self) -> &'static str {
-        QueueWorkArgs::DESCRIPTION
-    }
-    
-    fn help(&self) -> Option<&'static str> {
-        QueueWorkArgs::HELP
-    }
-}
-
-impl QueueWorkCommand {
-    pub fn new(args: QueueWorkArgs) -> Self {
-        Self { args }
-    }
-    
-    async fn verify_database_setup(&self) -> Result<(), CommandError> {
-        // Check if we have a database configuration
-        if !std::path::Path::new("Cargo.toml").exists() {
-            return Err(CommandError::ExecutionError(
-                "No Cargo.toml found. Make sure you're in an elif project directory.".to_string()
-            ));
-        }
-        
-        // Look for database configuration in common locations
-        let config_files = ["config/database.yml", "database.toml", ".env"];
-        let mut found_config = false;
-        
-        for config_file in &config_files {
-            if std::path::Path::new(config_file).exists() {
-                found_config = true;
-                if self.args.verbose {
-                    println!("📄 Found database config: {}", config_file);
-                }
-                break;
-            }
-        }
-        
-        if !found_config {
-            println!("⚠️  Warning: No database configuration found. Make sure you have:");
-            println!("   - config/database.yml, database.toml, or .env file");
-            println!("   - DATABASE_URL environment variable");
-        }
-        
-        Ok(())
-    }
-    
-    async fn start_workers(&self, queues: Vec<&str>) -> Result<(), CommandError> {
-        let mut processed_jobs = 0u32;
-        let sleep_duration = Duration::from_millis(self.args.sleep);
-        
-        println!("🚀 Workers started, listening for jobs...");
-        
-        // In a real implementation, this would:
-        // 1. Connect to the database
-        // 2. Query for pending jobs in the specified queues
-        // 3. Spawn worker tasks to process jobs
-        // 4. Handle job retries and failures
-        // 5. Update job status in the database
-        
-        loop {
-            // Simulate job processing
-            let jobs_found = self.check_for_jobs(&queues).await?;
-            
-            if jobs_found > 0 {
-                processed_jobs += jobs_found;
-                
-                if self.args.verbose {
-                    println!("✅ Processed {} jobs (total: {})", jobs_found, processed_jobs);
-                }
-                
-                // Check if we've reached the max job limit
-                if let Some(max) = self.args.max_jobs {
-                    if processed_jobs >= max {
-                        println!("🎯 Reached maximum job limit ({}), stopping", max);
-                        break;
-                    }
-                }
-            } else if self.args.stop_when_empty {
-                println!("📭 No jobs found and stop-when-empty is enabled, stopping");
-                break;
-            }
-            
-            // Handle Ctrl+C gracefully
-            tokio::select! {
-                _ = tokio::time::sleep(sleep_duration) => {
-                    // Continue processing
-                }
-                _ = tokio::signal::ctrl_c() => {
-                    println!("\n🛑 Received Ctrl+C, stopping workers gracefully...");
-                    self.shutdown_workers().await?;
-                    break;
-                }
-            }
-        }
-        
-        println!("✅ Queue worker stopped. Total jobs processed: {}", processed_jobs);
-        Ok(())
-    }
-    
-    async fn check_for_jobs(&self, queues: &[&str]) -> Result<u32, CommandError> {
-        // This is a placeholder implementation
-        // In a real application, this would:
-        // 1. Query the database for pending jobs
-        // 2. Return the number of jobs found and processed
-        
-        if self.args.verbose {
-            println!("🔍 Checking for jobs in queues: {:?}", queues);
-        }
-        
-        // Simulate finding jobs occasionally
-        use rand::Rng;
-        let mut rng = rand::thread_rng();
-        let found_jobs = if rng.gen_ratio(1, 10) { // 10% chance
-            rng.gen_range(1..=3)
-        } else {
-            0
-        };
-        
-        if found_jobs > 0 && self.args.verbose {
-            println!("🎯 Found {} jobs to process", found_jobs);
-        }
-        
-        Ok(found_jobs)
-    }
-    
-    async fn shutdown_workers(&self) -> Result<(), CommandError> {
-        println!("🔄 Gracefully shutting down workers...");
-        
-        // In a real implementation, this would:
-        // 1. Signal all worker tasks to stop
-        // 2. Wait for current jobs to complete
-        // 3. Clean up database connections
-        // 4. Save any necessary state
-        
-        tokio::time::sleep(Duration::from_millis(500)).await;
-        println!("✅ All workers stopped gracefully");
-        
-        Ok(())
-    }
-}
-
-/// Queue status command arguments
-#[derive(Args, Debug, Clone)]
-pub struct QueueStatusArgs {
-    /// Queue names to show status for
-    #[arg(long, short)]
-    pub queue: Option<String>,
-    
-    /// Show detailed job information
-    #[arg(long, short)]
-    pub detailed: bool,
-    
-    /// Refresh interval in seconds (0 for no refresh)
-    #[arg(long, short, default_value = "0")]
-    pub refresh: u64,
-}
-
-impl_command!(
-    QueueStatusArgs,
-    "queue:status",
-    "Show status of job queues",
-    "Display the current status of job queues including pending, processing, and failed jobs.\n\n\
-     Examples:\n\
-       elifrs queue:status\n\
-       elifrs queue:status --queue high\n\
-       elifrs queue:status --detailed --refresh 5"
-);
-
-/// Queue status command handler
-pub struct QueueStatusCommand {
-    pub args: QueueStatusArgs,
-}
-
-#[async_trait]
-impl CommandHandler for QueueStatusCommand {
-    async fn handle(&self) -> Result<(), CommandError> {
-        if self.args.refresh > 0 {
-            self.monitor_queues().await
-        } else {
-            self.show_status_once().await
-        }
-    }
-    
-    fn name(&self) -> &'static str {
-        QueueStatusArgs::NAME
-    }
-    
-    fn description(&self) -> &'static str {
-        QueueStatusArgs::DESCRIPTION  
-    }
-    
-    fn help(&self) -> Option<&'static str> {
-        QueueStatusArgs::HELP
-    }
-}
-
-impl QueueStatusCommand {
-    pub fn new(args: QueueStatusArgs) -> Self {
-        Self { args }
-    }
-    
-    async fn show_status_once(&self) -> Result<(), CommandError> {
-        println!("📊 Queue Status");
-        println!("═══════════════");
-        
-        // Mock queue data - in real implementation, query database
-        let queues = if let Some(queue_filter) = &self.args.queue {
-            vec![queue_filter.as_str()]
-        } else {
-            vec!["high", "default", "low", "mail"]
-        };
-        
-        for queue in queues {
-            self.show_queue_status(queue).await?;
-        }
-        
-        Ok(())
-    }
-    
-    async fn monitor_queues(&self) -> Result<(), CommandError> {
-        println!("🔄 Monitoring queues (refresh every {}s, Ctrl+C to stop)", self.args.refresh);
-        
-        let refresh_duration = Duration::from_secs(self.args.refresh);
-        
-        loop {
-            // Clear screen (basic version)
-            print!("\x1B[2J\x1B[1;1H");
-            
-            self.show_status_once().await?;
-            
-            println!("\nLast updated: {}", chrono::Utc::now().format("%Y-%m-%d %H:%M:%S UTC"));
-            
-            tokio::select! {
-                _ = tokio::time::sleep(refresh_duration) => {
-                    // Continue monitoring
-                }
-                _ = tokio::signal::ctrl_c() => {
-                    println!("\n🛑 Stopping queue monitor");
-                    break;
-                }
-            }
-        }
-        
-        Ok(())
-    }
-    
-    async fn show_queue_status(&self, queue_name: &str) -> Result<(), CommandError> {
-        // Mock data - replace with actual database queries
-        use rand::Rng;
-        let mut rng = rand::thread_rng();
-        
-        let pending = rng.gen_range(0..=50);
-        let processing = rng.gen_range(0..=5);
-        let completed = rng.gen_range(100..=1000);
-        let failed = rng.gen_range(0..=10);
-        
-        println!("\n📋 Queue: {}", queue_name);
-        println!("   ⏳ Pending:    {}", pending);
-        println!("   🔄 Processing: {}", processing);
-        println!("   ✅ Completed:  {}", completed);
-        println!("   ❌ Failed:     {}", failed);
-        
-        if self.args.detailed && (pending > 0 || processing > 0 || failed > 0) {
-            println!("   📝 Recent jobs:");
-            
-            if processing > 0 {
-                println!("      🔄 Job #12345 - send_email (started 2m ago)");
-            }
-            
-            if failed > 0 {
-                println!("      ❌ Job #12340 - process_payment (failed: timeout)");
-            }
-        }
-        
-        Ok(())
-    }
-}
-
-/// Create and run queue work command
-pub async fn work(args: QueueWorkArgs) -> Result<(), ElifError> {
-    let command = QueueWorkCommand::new(args);
-    command.handle().await.map_err(|e| {
-        ElifError::Codegen(format!("Queue work command failed: {}", e))
-    })
-}
-
-/// Create and run queue status command  
-pub async fn status(args: QueueStatusArgs) -> Result<(), ElifError> {
-    let command = QueueStatusCommand::new(args);
-    command.handle().await.map_err(|e| {
-        ElifError::Codegen(format!("Queue status command failed: {}", e))
-    })
-}
-
 /// Schedule run command arguments
 #[derive(Args, Debug, Clone)]
 pub struct ScheduleRunArgs {
