+++ crates/cli/src/commands/queue/queue_status.rs
@@ -1,0 +1,150 @@
+use std::time::Duration;
+use clap::Args;
+use elif_core::ElifError;
+use crate::command_system::{CommandHandler, CommandError, CommandDefinition, impl_command};
+use async_trait::async_trait;
+
+/// Queue status command arguments
+#[derive(Args, Debug, Clone)]
+pub struct QueueStatusArgs {
+    /// Queue names to show status for
+    #[arg(long, short)]
+    pub queue: Option<String>,
+    
+    /// Show detailed job information
+    #[arg(long, short)]
+    pub detailed: bool,
+    
+    /// Refresh interval in seconds (0 for no refresh)
+    #[arg(long, short, default_value = "0")]
+    pub refresh: u64,
+}
+
+impl_command!(
+    QueueStatusArgs,
+    "queue:status",
+    "Show status of job queues",
+    "Display the current status of job queues including pending, processing, and failed jobs.\n\n\
+     Examples:\n\
+       elifrs queue:status\n\
+       elifrs queue:status --queue high\n\
+       elifrs queue:status --detailed --refresh 5"
+);
+
+/// Queue status command handler
+pub struct QueueStatusCommand {
+    pub args: QueueStatusArgs,
+}
+
+#[async_trait]
+impl CommandHandler for QueueStatusCommand {
+    async fn handle(&self) -> Result<(), CommandError> {
+        if self.args.refresh > 0 {
+            self.monitor_queues().await
+        } else {
+            self.show_status_once().await
+        }
+    }
+    
+    fn name(&self) -> &'static str {
+        QueueStatusArgs::NAME
+    }
+    
+    fn description(&self) -> &'static str {
+        QueueStatusArgs::DESCRIPTION  
+    }
+    
+    fn help(&self) -> Option<&'static str> {
+        QueueStatusArgs::HELP
+    }
+}
+
+impl QueueStatusCommand {
+    pub fn new(args: QueueStatusArgs) -> Self {
+        Self { args }
+    }
+    
+    async fn show_status_once(&self) -> Result<(), CommandError> {
+        println!("📊 Queue Status");
+        println!("═══════════════");
+        
+        // Mock queue data - in real implementation, query database
+        let queues = if let Some(queue_filter) = &self.args.queue {
+            vec![queue_filter.as_str()]
+        } else {
+            vec!["high", "default", "low", "mail"]
+        };
+        
+        for queue in queues {
+            self.show_queue_status(queue).await?;
+        }
+        
+        Ok(())
+    }
+    
+    async fn monitor_queues(&self) -> Result<(), CommandError> {
+        println!("🔄 Monitoring queues (refresh every {}s, Ctrl+C to stop)", self.args.refresh);
+        
+        let refresh_duration = Duration::from_secs(self.args.refresh);
+        
+        loop {
+            // Clear screen (basic version)
+            print!("\x1B[2J\x1B[1;1H");
+            
+            self.show_status_once().await?;
+            
+            println!("\nLast updated: {}", chrono::Utc::now().format("%Y-%m-%d %H:%M:%S UTC"));
+            
+            tokio::select! {
+                _ = tokio::time::sleep(refresh_duration) => {
+                    // Continue monitoring
+                }
+                _ = tokio::signal::ctrl_c() => {
+                    println!("\n🛑 Stopping queue monitor");
+                    break;
+                }
+            }
+        }
+        
+        Ok(())
+    }
+    
+    async fn show_queue_status(&self, queue_name: &str) -> Result<(), CommandError> {
+        // Mock data - replace with actual database queries
+        use rand::Rng;
+        let mut rng = rand::thread_rng();
+        
+        let pending = rng.gen_range(0..=50);
+        let processing = rng.gen_range(0..=5);
+        let completed = rng.gen_range(100..=1000);
+        let failed = rng.gen_range(0..=10);
+        
+        println!("\n📋 Queue: {}", queue_name);
+        println!("   ⏳ Pending:    {}", pending);
+        println!("   🔄 Processing: {}", processing);
+        println!("   ✅ Completed:  {}", completed);
+        println!("   ❌ Failed:     {}", failed);
+        
+        if self.args.detailed && (pending > 0 || processing > 0 || failed > 0) {
+            println!("   📝 Recent jobs:");
+            
+            if processing > 0 {
+                println!("      🔄 Job #12345 - send_email (started 2m ago)");
+            }
+            
+            if failed > 0 {
+                println!("      ❌ Job #12340 - process_payment (failed: timeout)");
+            }
+        }
+        
+        Ok(())
+    }
+}
+
+/// Create and run queue status command  
+pub async fn status(args: QueueStatusArgs) -> Result<(), ElifError> {
+    let command = QueueStatusCommand::new(args);
+    command.handle().await.map_err(|e| {
+        ElifError::Codegen(format!("Queue status command failed: {}", e))
+    })
+}
